Updated Rector to commit 17cfa9f376b5b4a00a34e3b09ac48e7bd812be3f
https://github.com/rectorphp/rector-src/commit/17cfa9f376b5b4a00a34e3b09ac48e7bd812be3f Update to nikic/php-parser 4.13.0 (#904)

+++ rules/MysqlToMysqli/Rector/FuncCall/MysqlFuncCallToMysqliRector.php
@@ -44,7 +44,7 @@ CODE_SAMPLE
     /**
      * @param FuncCall $node
      */
-    public function refactor(\PhpParser\Node $node) : \PhpParser\Node\Expr\FuncCall
+    public function refactor(\PhpParser\Node $node) : ?\PhpParser\Node\Expr\FuncCall
     {
         if ($this->isName($node, 'mysql_create_db')) {
             return $this->processMysqlCreateDb($node);
@@ -56,25 +56,37 @@ CODE_SAMPLE
             $node->name = new \PhpParser\Node\Name(self::MYSQLI_QUERY);
             $node->args[0] = new \PhpParser\Node\Arg(new \PhpParser\Node\Scalar\String_('SHOW DATABASES'));
         }
-        if ($this->isName($node, 'mysql_list_fields')) {
+        if ($this->isName($node, 'mysql_list_fields') && $node->args[0] instanceof \PhpParser\Node\Arg && $node->args[1] instanceof \PhpParser\Node\Arg) {
             $node->name = new \PhpParser\Node\Name(self::MYSQLI_QUERY);
             $node->args[0]->value = $this->joinStringWithNode('SHOW COLUMNS FROM', $node->args[1]->value);
             unset($node->args[1]);
         }
-        if ($this->isName($node, 'mysql_list_tables')) {
+        if ($this->isName($node, 'mysql_list_tables') && $node->args[0] instanceof \PhpParser\Node\Arg) {
             $node->name = new \PhpParser\Node\Name(self::MYSQLI_QUERY);
             $node->args[0]->value = $this->joinStringWithNode('SHOW TABLES FROM', $node->args[0]->value);
         }
         return $node;
     }
-    private function processMysqlCreateDb(\PhpParser\Node\Expr\FuncCall $funcCall) : \PhpParser\Node\Expr\FuncCall
+    private function processMysqlCreateDb(\PhpParser\Node\Expr\FuncCall $funcCall) : ?\PhpParser\Node\Expr\FuncCall
     {
+        if (!isset($funcCall->args[0])) {
+            return null;
+        }
+        if (!$funcCall->args[0] instanceof \PhpParser\Node\Arg) {
+            return null;
+        }
         $funcCall->name = new \PhpParser\Node\Name(self::MYSQLI_QUERY);
         $funcCall->args[0]->value = $this->joinStringWithNode('CREATE DATABASE', $funcCall->args[0]->value);
         return $funcCall;
     }
-    private function processMysqlDropDb(\PhpParser\Node\Expr\FuncCall $funcCall) : \PhpParser\Node\Expr\FuncCall
+    private function processMysqlDropDb(\PhpParser\Node\Expr\FuncCall $funcCall) : ?\PhpParser\Node\Expr\FuncCall
     {
+        if (!isset($funcCall->args[0])) {
+            return null;
+        }
+        if (!$funcCall->args[0] instanceof \PhpParser\Node\Arg) {
+            return null;
+        }
         $funcCall->name = new \PhpParser\Node\Name(self::MYSQLI_QUERY);
         $funcCall->args[0]->value = $this->joinStringWithNode('DROP DATABASE', $funcCall->args[0]->value);
         return $funcCall;
